requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
plotly==5.18.0
scikit-learn==1.4.0
gradio==4.19.2
python-dotenv==1.0.1
dvc==3.50.2
pytest==8.0.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0
matplotlib>=3.6.0
seaborn>=0.12.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
keras>=2.12.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.1.0
optuna>=3.1.0
mlflow>=2.3.0
tqdm>=4.62.0
ydata-profiling>=4.1.0
sweetviz>=2.1.3
great-expectations>=0.15.0
pandera>=0.7.0
pyyaml>=5.4.1
tqdm>=4.62.0 